+++ js/common.js
@@ -337,23 +337,18 @@
 	 * @example {Object} $e_parsePredefinedConstants("maxX")
 	 */
 	function $e_parsePredefinedConstants(value) {
-		var canvasSize = $_eseecode.whiteboard.offsetWidth;
 		if (value == "minX") {
-			value = -$_eseecode.coordinates.position.x;
+			value = $e_system2userCoords({x: 0, y: 0}).x;
 		} else if (value == "maxX") {
-			value = canvasSize-$_eseecode.coordinates.position.x;
+			value = $e_system2userCoords({x: getLayerWidth(), y: getLayerHeight()}).x;
 		} else if (value == "minY") {
-			value = -$_eseecode.coordinates.position.y;
+			value = $e_system2userCoords({x: 0, y: 0}).y;
 		} else if (value == "maxY") {
-			value = canvasSize-$_eseecode.coordinates.position.y;
+			value = $e_system2userCoords({x: getLayerWidth(), y: getLayerHeight()}).y;
 		} else if (value == "centerX") {
-			value = $e_system2userCoords({x: canvasSize/2, y: canvasSize/2}).x;
+			value = $e_system2userCoords({x: getLayerWidth()/2, y: getLayerHeight()/2}).x;
 		} else if (value == "centerY") {
-			value = $e_system2userCoords({x: canvasSize/2, y: canvasSize/2}).y;
-		} else if (value == "originX") {
-			value = $e_user2systemCoords({x: 0, y: 0}).x;
-		} else if (value == "originY") {
-			value = $e_user2systemCoords({x: 0, y: 0}).y;
+			value = $e_system2userCoords({x: getLayerWidth()/2, y: getLayerHeight()/2}).y;
 		}
 		return value;
 	}
@@ -367,8 +362,8 @@
 	 */
 	function $e_user2systemCoords(pos) {
 		var value = {};
-		value.x = pos.x*$_eseecode.coordinates.scale.x+$_eseecode.coordinates.position.x;
-		value.y = pos.y*$_eseecode.coordinates.scale.y+$_eseecode.coordinates.position.y;
+		value.x = pos.x*$_eseecode.coordinates.scale.x+$_eseecode.coordinates.origin.x;
+		value.y = pos.y*$_eseecode.coordinates.scale.y+$_eseecode.coordinates.origin.y;
 		return value;
 	}
 	
@@ -381,8 +376,8 @@
 	 */
 	function $e_system2userCoords(pos) {
 		var value = {};
-		value.x = (pos.x-$_eseecode.coordinates.position.x)/$_eseecode.coordinates.scale.x;
-		value.y = (pos.y-$_eseecode.coordinates.position.y)/$_eseecode.coordinates.scale.y;
+		value.x = (pos.x-$_eseecode.coordinates.origin.x)/$_eseecode.coordinates.scale.x;
+		value.y = (pos.y-$_eseecode.coordinates.origin.y)/$_eseecode.coordinates.scale.y;
 		return value;
 	}